Adding Controllers
Started to implement MVC pattern in our project.
Created the controller files and moved the controller logic from the routes files to their respective files.
refactored the app.js to work with this new reorganization

closes #63
