+++ controllers/errors.js
@@ -1,0 +1,3 @@
+exports.get404Error = (req, res, next) => {
+    res.status(404).render('404', {pageTitle: 'Page Not Found', path: ""});
+};
