+++ controllers/products.js
@@ -1,0 +1,29 @@
+const products = [];
+
+exports.getAddProduct = (req, res, next) => {
+  res.render("add-product", {
+    pageTitle: "Add Product",
+    path: "/admin/add-product",
+    formCSS: true,
+    productCSS
+: true,
+    activeAddProduct: true,
+  });
+};
+
+exports.postAddProduct = (req, res, next) => {
+    products.push({product: req.body.product});
+    res.redirect('/');
+};
+
+exports.getProducts = (req, res, next) => {
+    res.render("shop", {
+      prods: products,
+      pageTitle: "Shop",
+      path: "/",
+      hasProducts: products.length > 0,
+      activeShop: true,
+      productCSS: true,
+    });
+};
+
