+++ routes/admin.js
@@ -2,26 +2,13 @@ const path = require('path');
 
 const express = require('express');
 
-const rootDir = require('../utils/path');
+const productsController = require('../controllers/products');
 
 const router = express.Router();
 
-const products = [];
 
-router.get('/add-product', (req, res, next) => {
-    res.render('add-product', {
-                                pageTitle: 'Add Product', 
-                                path: '/admin/add-product',
-                                formCSS: true,
-                                productCSS: true,
-                                activeAddProduct: true
-                            });
-});
+router.get('/add-product', productsController.getAddProduct);
 
-router.post('/add-product', (req, res, next) => {
-    products.push({product: req.body.product});
-    res.redirect('/');
-});
+router.post('/add-product', productsController.postAddProduct);
 
-exports.routes = router;
-exports.products = products;
+module.exports = router;
