+++ routes/shop.js
@@ -1,22 +1,11 @@
-const path = require('path');
+const path = require("path");
 
-const express = require('express');
+const express = require("express");
 
-const rootDir = require('../utils/path');
-const adminData = require('./admin');
+const productsController = require("../controllers/products");
 
 const router = express.Router();
 
-router.get('/', (req, res, next) => {
-    const products  = adminData.products;
-    res.render('shop',{
-                prods: products, 
-                pageTitle: 'Shop', 
-                path: '/', 
-                hasProducts: products.length > 0,
-                activeShop: true,
-                productCSS: true,
-            });
-});
+router.get("/", productsController.getProducts);
 
 module.exports = router;
